Fixed languages. Every line is composed with this parameters: Continent/Region:locale:language
The "language" field refer to the language list located in the file all_kde_langpacks. If a language is not found in this file but is written in timezones, tribe show a blank box. The user can select the second value, most of the cases is en_gb

pt_br -> pt_BR
zh_cn -> zh_CN
zh_tw -> zh_TW
jp -> ja (fix for Tokyo)